+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Please, check out guidelines: https://keepachangelog.com/en/1.0.0/
 
 ### Fixed
 
+- Fix static products false positive lint warning by https://github.com/tuist/tuist/pull/981 @kwridan.
+
 ### Changed
 
 ## 1.2.0